Do basic optimization passes.
The expected use of the modules in this repo is to link them at the IR
level with other code, so the optimizations used here don't really
achieve much except giving a sense of how the code might look after
optimizations are applied.

Since this is primarily a developer tool, the optimized bitcode is
converted by the Makefile back to LLVM IR for human inspection.