waitqueue_wait_hipri and waitqueue_wait_lopri
Priority here is implemented just by which end of the list the new item
is added on. High-priority items are added at the head of the queue,
causing them to cut in line in front of everyone else, while low-priority
items are added to the tail of the queue, waiting for all existing items
to be processed.